feat(map): pull out map.tsx functionality into multiple files
